+++ CHANGELOG.md
@@ -7,12 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-<<<<<<< HEAD
-=======
 ### Changed
+- Generate metadata about service classes during FIR using `FirDeclarationDataRegistry` so that IR can read them directly instead of reparsing the `@AutoService` annotations and classes.
 - Remove mirror classes during IR generation
 
->>>>>>> b6bec2a (Strip mirrors with IR)
 ## [0.1.0]
 
 ### Added